+++ wemake_python_styleguide/checker.py
@@ -77,5 +77,4 @@ class Checker(object):
             visiter.visit(self.tree)
 
             for error in visiter.errors:
-                lineno, col_offset, message = error.node_items()
-                yield lineno, col_offset, message, type(self)
+                yield (*error.node_items(), type(self))